All modifications done with gettimeofday

+++ file.h
@@ -13,7 +13,7 @@ void write_PROC_EXIT(int exit_code);
 void write_FILE_MODF(char*old_mode,char* new_mode,char*file_name);
 void write_SIGNAL_RECV(char *signal);
 void write_SIGNAL_SENT(char *signal,pid_t target_pid);
-double calculate_time();
+long long calculate_time();
 //void eventHandler(int code, int argc, char *argv[], char*reg,double time_taken);
 
 void getfd(char*reg);
